requirements.txt
requests
tqdm
progressbar2
py7zr
psycopg2-binary
more-itertools